extract cluster extract to a separate module
